requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
pytz>=2023.3
python-dateutil>=2.8.2
asyncpg>=0.29.0
gspread>=5.0.0
google-auth>=2.22.0
oauth2client>=4.1.3
psutil>=5.9.8